tests/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
flask>=2.3.0
PyYAML>=5.4
# E2E（可选）: pip install pytest-playwright && playwright install chromium
# pytest-playwright>=0.4.0
